9.2 Exercise calculator
Write a function calculateExercises that calculates the average time of daily exercise hours and compares it to the target amount of daily hours and returns an object that includes the following values:

-the number of days
-the number of training days
-the original target value
-the calculated average time
-boolean value describing if the target was reached
a rating between the numbers 1-3 that tells how well the hours are met. You can decide on the metric on your own.
-a text value explaining the rating