Exercise 9.14
Create a new Create React App with TypeScript, and set up eslint for the project similarly to how we just did.

The whole app is now in one component. That is not what we want, so refactor the code so that it consists of three components: Header, Content and Total. All data is still kept in the App component, which passes all necessary data to each component as props. Be sure to add type declarations for each component's props!