9.20: patientor, step5
Extend a patient's page in the frontend to list the date, description and diagnose codes of the patient's entries.

You can use the same type definition for an Entry in the frontend. For these exercises, it is enough to just copy/paste the definitions from the backend to the frontend.